pyproject.toml
[tool.poetry]
name = "ns-controller"
version = "0.1.0"
description = "Nintendo Switch Pro Controller emulator over USB with TCP server"
authors = ["Dylan Meadows <dylmeadows@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.120.3"}
loguru = "^0.7.3"
requests = "^2.32.5"
click = "^8.3.0"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.14.2"
opencv-python = "^4.12.0.88"
pytesseract = "^0.3.13"
pillow = "^11.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
exclude = [".venv", "__pycache__", "build", "dist"]
line-length = 120
indent-width = 4
[tool.ruff.format]
quote-style = "double"
[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "C90"]
[tool.ruff.lint.mccabe]
max-complexity = 20

